requirements.txt
streamlit
pandas
plotly
sqlalchemy
streamlit-authenticator
PyYAML
bcrypt
fpdf2
holidays
openpyxl
psycopg2-binary
